added bar graphics

+++ KotlinProject/app/src/main/java/jefemayoneso/compi1prac1/Backend/ParserActions/PieGraphicActioner.java
@@ -11,7 +11,7 @@ public class PieGraphicActioner {
 
     public boolean titleDeclaration(String title, PieGraphic pieGraph) {
         // add message to errors and set error to object
-        pieGraph.setTitleDecl((short) (pieGraph.getTitleDecl() + 1)); // title decs ++
+        pieGraph.setTitleDecl(pieGraph.getTitleDecl() + 1); // title decs ++
         if (pieGraph.getTitleDecl() <= 1) {
             // save info for execution
             pieGraph.setTitle(title.substring(1, title.length() - 1)); // save info
@@ -22,7 +22,7 @@ public class PieGraphicActioner {
 
     public boolean typeDeclaration(int type, PieGraphic pieGraphic) {
         // save data
-        pieGraphic.setPieTypeDecl((short) (pieGraphic.getPieTypeDecl() + 1));
+        pieGraphic.setPieTypeDecl(pieGraphic.getPieTypeDecl() + 1);
         if (pieGraphic.getPieTypeDecl() <= 1) {
             // save info for execution
             String typeS = type == 0 ? "Cantidad" : "Porcentaje";
@@ -35,7 +35,7 @@ public class PieGraphicActioner {
     boolean tagsDeclaration(String items, PieGraphic pieGraphic) {
         // split item with \n for files
         String[] itemsSplit = items.split("\n");
-        pieGraphic.setTagsDecl((short) (pieGraphic.getTagsDecl() + 1));
+        pieGraphic.setTagsDecl(pieGraphic.getTagsDecl() + 1);
         // validate lenght > 0
         if (pieGraphic.getTagsDecl() <= 1 && itemsSplit.length > 0) {
             pieGraphic.setTags(itemsSplit);
@@ -51,7 +51,7 @@ public class PieGraphicActioner {
         for (int i = 0; i < itemsSplit.length; i++) {
             numbers[i] = Double.valueOf(itemsSplit[i]);
         }
-        pieGraphic.setValuesDecl((short) (pieGraphic.getValuesDecl() + 1));
+        pieGraphic.setValuesDecl(pieGraphic.getValuesDecl() + 1);
         if (pieGraphic.getValuesDecl() <= 1 && numbers.length > 0) {
             pieGraphic.setValues(numbers);
             return true;
@@ -60,7 +60,7 @@ public class PieGraphicActioner {
     }
 
     boolean totalDeclaration(Double number, PieGraphic pieGraphic) {
-        pieGraphic.setTotalDecl((short) (pieGraphic.getTotalDecl() + 1));
+        pieGraphic.setTotalDecl(pieGraphic.getTotalDecl() + 1);
         if (pieGraphic.getTotalDecl() <= 1) {
             pieGraphic.setTotal(number);
             return true;
@@ -68,8 +68,8 @@ public class PieGraphicActioner {
         return false;
     }
 
-    boolean mergeDeclaration(ArrayList<short[]> mergeCoordenates, PieGraphic pieGraphic) {
-        pieGraphic.setMergeDecl((short) (pieGraphic.getMergeDecl() + 1));
+    boolean mergeDeclaration(ArrayList<int[]> mergeCoordenates, PieGraphic pieGraphic) {
+        pieGraphic.setMergeDecl(pieGraphic.getMergeDecl() + 1);
         if (pieGraphic.getMergeDecl() <= 1 && mergeCoordenates.size() > 0) {
             pieGraphic.setMergeItems(mergeCoordenates);
             return true;
@@ -78,7 +78,7 @@ public class PieGraphicActioner {
     }
 
     boolean extraDeclaration(String extra, PieGraphic pieGraphic) {
-        pieGraphic.setExtraDecl((short) (pieGraphic.getExtraDecl() + 1));
+        pieGraphic.setExtraDecl(pieGraphic.getExtraDecl() + 1);
         if (pieGraphic.getExtraDecl() <= 1) {
             pieGraphic.setExtra(extra);
             return true;
